No Edgeproblem between the blocks

+++ Assets/Resources/Scripts/Block.cs
@@ -5,17 +5,17 @@ public class Block{
 
     public static Block[] blockData =
     {
-        new Block(-1, -1, true),                        //Air
+        new Block(-1f, -1f, true),                        //Air
         new Block(0, 0),                                //Stone
-        new Block(1, 0),                                //Dirt
-        new Block(2, 0, 2, 0, 3, 0, 3, 0, 2, 0, 2, 0),  //Wood
-        new Block(4, 0, true),                          //Leaves
-        new Block(6, 0, 6, 0, 1, 0, 5, 0, 6, 0, 6, 0),  //Grass
-        new Block(7, 0, true),                          //Glass
+        new Block(1f, 0),                                //Dirt
+        new Block(2f, 0, 2f, 0, 3f, 0, 3f, 0, 2f, 0, 2f, 0),  //Wood
+        new Block(4f, 0, true),                          //Leaves
+        new Block(6f, 0, 6f, 0, 1f, 0, 5f, 0, 6f, 0, 6f, 0),  //Grass
+        new Block(7f, 0, true),                          //Glass
     };
 
     //UV-Coordinates for every Face
-    public int xLeft, yLeft, 
+    public float xLeft, yLeft, 
                xRight, yRight, 
                xBottom, yBottom, 
                xTop, yTop, 
@@ -24,18 +24,18 @@ public class Block{
     public bool isTransparent = false;
 
     //Creates a block with the Textur at x/y
-	public Block(int x, int y) : this(x, y, x, y, x, y, x, y, x, y, x, y)
+	public Block(float x, float y) : this(x, y, x, y, x, y, x, y, x, y, x, y)
     {
 
     }
 
     //Creates a block with the Textur at x/y
-    public Block(int x, int y, bool isTransparent) : this(x, y, x, y, x, y, x, y, x, y, x, y)
+    public Block(float x, float y, bool isTransparent) : this(x, y, x, y, x, y, x, y, x, y, x, y)
     {
         this.isTransparent = isTransparent;
     }
 
-    public Block(int xLeft, int yLeft, int xRight, int yRight, int xBottom, int yBottom, int xTop, int yTop, int xFront, int yFront, int xBack, int yBack)
+    public Block(float xLeft, float yLeft, float xRight, float yRight, float xBottom, float yBottom, float xTop, float yTop, float xFront, float yFront, float xBack, float yBack)
     {
         this.xLeft = xLeft;
         this.yLeft = yLeft;
